cleanup command packets for 1.20.5 (#5)
